backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Deep Learning & ML
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
facenet-pytorch==2.6.0
transformers==5.3.0
numpy==1.26.4

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Audio Processing
librosa==0.10.1

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25

# Utilities
loguru==0.7.2
psutil==5.9.7
python-magic==0.4.27
